pipeline/pyproject.toml
[project]
name = "cookbook-pipeline"
version = "0.1.0"
description = "Extract structured recipes from the India Cookbook PDF"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-mock>=3.12.0", "ruff>=0.4.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
